% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_scopus.R
\name{read_scopus}
\alias{read_scopus}
\title{Read Scopus collection}
\usage{
read_scopus(file, TC_min = 0, TC_year_min = 0, PY_max = Inf,
  PY_min = 0, n_max = Inf, type = NULL, fields = NULL,
  exclude = NULL)
}
\arguments{
\item{TC_min}{Minimal number of citations for the publication to be included.}

\item{TC_year_min}{Minimal number of annual citations for the publication to be included.}

\item{PY_max}{Latest year of publication to be included.}

\item{PY_min}{Earliest year of publication to be included.}

\item{n_max}{Only include the first n publications per file.}

\item{type}{Latest year of publication to be included.}

\item{fields}{Fields to include. Either type or fields has to be specified.}

\item{exclude}{Fields to exclude}

\item{file_list}{The paths to the downloaded Scopus csv files.}

\item{type}{Type of report, either "complete" or "reduced". Either type or fields has to be specified.}
}
\description{
Reads and parses a collection of multiple csv files from Scopus (wrapper around read_scopus).
}
\details{
With this function, multiple Scopus csv downloads can be read and parsed.
}
