% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_scopus.R
\name{read_scopus_collection}
\alias{read_scopus_collection}
\title{read_scopus_csv_collection()}
\usage{
read_scopus_collection(file_list, TC_min = 0, TC_year_min = 0,
  PY_min = 0, PY_max = Inf, n_max = Inf, type = NULL,
  fields = NULL, exclude = NULL)
}
\arguments{
\item{file_list}{The paths to the downloaded Scopus csv files.}
}
\description{
Reads and parses a collection of multiple csv files from Scopus (wrapper around read_scopus).
}
\details{
XXXXXXXXXX
}
