% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_scopus.R
\name{scopus_document_ID}
\alias{scopus_document_ID}
\title{Search Scopus by ID}
\usage{
scopus_document_ID(ID, idtype = "eid", type = "abstract",
  view = "FULL", scopus_key, start = 1, t_limit = 6)
}
\arguments{
\item{ID}{An scopus identifier (or a vector of many). Can be either the EID or DOI of the publication.}

\item{idtype}{The type of ID provided. Either "eid" or "doi".}

\item{start}{Number of ID in the vector where the iteration starts.}

\item{t_limit}{The time in seconds untill a new iteration starts. Limited by Scopus.}

\item{Your}{scopus API key.}

\item{content}{If the "complete" or "reduced" publication information should be extracted.}
}
\description{
Retrieves scopus documents via ID search one-by-one in different views
}
\details{
Retrieves scopus documents via ID search one-by-one in different views
}
