% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_key_management.R
\name{scopus_key}
\alias{scopus_key}
\title{Set or get a Scopus user key.}
\usage{
scopus_key(force = FALSE)
}
\arguments{
\item{force}{Overwrite existing key? If TRUE, the user will be prompted to
enter a user key even if one has already been entered before.}
}
\description{
Reads (or writes) the user key to an environment variable.
}
\details{
Request a user key from Scopus \url{https://dev.elsevier.com/}
in order to use the API. This key gets appended to GET queries via the
\code{user_key} parameter. The key will be saved so you don't have to enter
it repeatedly.
}
