% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_scopus.R
\name{scopus_search_ID}
\alias{scopus_search_ID}
\title{Scopus bulk download by ID}
\usage{
scopus_search_ID(ID, idtype, datatype = "application/json", scopus_key,
  content = "complete", start = 0, retCount = 200, t_limit = 6,
  outfile)
}
\arguments{
\item{ID}{An scopus identifier (or a vector of many). Can be either the EID or DOI of the publication.}

\item{idtype}{The type of ID provided. Either "eid" or "doi".}

\item{datatype}{Up to now only "application/json" supported.}

\item{content}{If the "complete" or "reduced" publication information should be extracted.}

\item{start}{Number of ID in the vector where the iteration starts.}

\item{retCount}{How many publications should be extracted per iteration. Up to now limited by Scopus.}

\item{t_limit}{The time in seconds untill a new iteration starts. Limited by Scopus.}

\item{outfile}{Path of the file to be saved.}

\item{Your}{scopus API key.}
}
\description{
Retrieves scopus articles via ID search.
}
\details{
Retrieves scopus articles (abstracts) via ID search. Therefore, up to 25-200 entries can be recieved for 1 request.
}
