% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_general.R
\name{weight_jaccard}
\alias{weight_jaccard}
\title{Jaccard weighting of edge-strenght}
\usage{
weight_jaccard(i, j, w = 1)
}
\arguments{
\item{i}{The outgoing node in an edgelist.}

\item{j}{The outgoing node in an edgelist.}

\item{w}{The weight of the edge between i and j.}
}
\description{
Jaccard weighting of edge-strenght in an edge list
}
\details{
Takes a weighted edge-list and adjusts the weight by dividing the two nodes
union by their intercept.
}
